Set up database testing environment

+++ db/migrations/02_create_spaces_table.sql
@@ -1,1 +1,1 @@
-CREATE TABLE spaces(id SERIAL PRIMARY KEY, title VARCHAR(60), price INTEGER(8), description VARCHAR(240));
+CREATE TABLE spaces(id SERIAL PRIMARY KEY, title VARCHAR(60), price INT, description VARCHAR(240));
